Hotspot Manager Settings and Logging update!
Added an icon in the taskbar's tray that shows when the task is running.

Right clicking this will end the task, there's also a "settings" option so you can change how often the task checks whether hotspot tethering is on. This is in ms!!! Default is 15s (15000ms). Settings are stored in local appdata. A log is stored in TEMP.